Refactor toast handler imports and update axios instance base URL handling
